example updated & minor fixes
