+++ example/android/settings.gradle
@@ -1,5 +1,5 @@
 rootProject.name = 'example'
-
-include ':app'
 include ':react-native-sensitive-info'
 project(':react-native-sensitive-info').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-sensitive-info/android')
+
+include ':app'
