+++ example/app.json
@@ -1,0 +1,4 @@
+{
+  "name": "example",
+  "displayName": "example"
+}
