+++ example/package.json
@@ -5,8 +5,8 @@
     "start": "node node_modules/react-native/local-cli/cli.js start"
   },
   "dependencies": {
-    "react": "~15.1.0",
-    "react-native": "^0.27.0",
+    "react": "^16.0.0-alpha.6",
+    "react-native": "^0.44.0",
     "react-native-sensitive-info": "file:../"
   }
 }
